fix define-flow-type and no-use-before-define compat
By using `globalScope.__define` we were adding a reference to the indent
node in `identifiers`. "no-use-before-define" tests for `identifiers.length`
to determine whether to check for variable usage. the logic there expected
that if there was a node in `identifiers` then it must also have a `def`.
But we don't have any defs for flow types because we treat them as explicit
globals.

`__defineGeneric` is called by `__define` with the node to add to
`identifiers`. By using `__defineGeneric ` and skipping the 4th arg,
nothing gets added to `identifiers`.